pyproject.toml
[tool.poetry]
name = "vk-chat-bot"
version = "0.1.0"
description = "VK chatbot demonstrating showcase to customers"
authors = ["Andrey_Fominykh <meriadok2003@mail.ru>"]
packages = [
    { include = "vk_bot" }
]

[tool.poetry.dependencies]
python = "^3.8"
psycopg2-binary = "^2.9.3"
peewee = "^3.15.2"
python-dotenv = "^0.20.0"
vk-api = "^11.9.9"
prompt = "^0.4.1"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
pytest = "^7.1.3"
pytest-cov = "^3.0.0"

[tool.poetry.scripts]
vk_bot = "vk_bot.scripts.vk_bot:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
